breaking: add ServerObjectManager for object spawning
