+++ Assets/Mirror/Runtime/ServerObjectManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b750203ddb2d9f84abd799a5c2c32cb6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
